Added full namespace automatically to Operation names.
This is to avoid situations where queues are defined in multiple implementations of the same base class, and the operation uses the same name but is different in the namespace. With this tweak the inferred `OperationId` is different from any other if they are defined in different types or namespaces.